#1091: Provide more dynamic services command and put it into its own file for easier access from GUI
